CFRA-API/cfra_server_files/api1/requirements.txt
pandas
requests
futures
mysql-connector-python